+++ CHANGELOG.md
@@ -6,10 +6,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-### Changed
+## [0.28.0] - 2020-05-04
+
+### Added
 - Add helpers to split transactions (`models.transaction.createSplit`, `models.transaction.deleteSplit`, `models.transaction.updateSplit`)
-- *Breaking*: renamed `StandingOrderReoccurenceType` to `StandingOrderReoccurrenceType`
 - Add method to fetch account stats (`models.account.getStats`)
+
+### Changed
+- *Breaking*: renamed `StandingOrderReoccurenceType` to `StandingOrderReoccurrenceType`
 
 ## [0.27.0] - 2020-04-20
 ### Added